requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pdfplumber==0.10.3
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic>=2.8.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pillow==10.1.0
pytesseract==0.3.10
